refactor(desktop): extract useActionPlan from App.tsx (F5.6)
The inbox action-plan / analyzer subsystem moves into useActionPlan:
running the AI analysis (or deterministic rules), the plan-tree nav
(planNodes/planNav + mirror refs), applying a category (archive/trash/
label/move/prompt), per-email preview, analyzer-rule CRUD, and the
analyzer-prompt view. Verbatim logic move; plan/rules state stays in App
and is passed in.

App.tsx 3,066 -> 2,734; useActionPlan 409 (< 500). unit 61 / e2e 36 green
(actionplan + rules specs cover it), tsc + build clean.
